Split the float, math, bool and string bridge fns out of bridge.rs
